Move UnknownRepositoryCollectionException to package org.molgenis.data
