+++ molgenis-data/src/test/java/org/molgenis/data/meta/EntityTypeRepositoryDecoratorTest.java
@@ -8,10 +8,10 @@ import org.mockito.quality.Strictness;
 import org.molgenis.data.DataService;
 import org.molgenis.data.Repository;
 import org.molgenis.data.RepositoryCollection;
+import org.molgenis.data.UnknownRepositoryCollectionException;
 import org.molgenis.data.meta.model.Attribute;
 import org.molgenis.data.meta.model.EntityType;
 import org.molgenis.data.meta.system.SystemEntityTypeRegistry;
-import org.molgenis.data.transaction.UnknownRepositoryCollectionException;
 import org.molgenis.test.AbstractMockitoTest;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
